+++ src/pages/BaseLayout/index.tsx
@@ -1,0 +1,4 @@
+const BaseLayout = () => {
+  return <div>안녕</div>;
+};
+export default BaseLayout;
